Move render_own_name back to Google Fonts profile
Renaming it to googlefonts/render_own_name
Migrating from Universal back to Google Fonts

based on feedback by @khaledhosny

(issue #4944)
