+++ Lib/fontbakery/legacy_checkids.py
@@ -130,6 +130,7 @@ renaming_map = {
     "com.google.fonts/check/old_ttfautohint":                               "googlefonts/old_ttfautohint",
     "com.google.fonts/check/production_encoded_glyphs":                     "googlefonts/production_encoded_glyphs",
     "com.google.fonts/check/production_glyphs_similarity":                  "googlefonts/production_glyphs_similarity",
+    "com.google.fonts/check/render_own_name":                               "googlefonts/render_own_name",
     "com.google.fonts/check/repo/dirname_matches_nameid_1":                 "googlefonts/repo/dirname_matches_nameid_1",
     "com.google.fonts/check/repo/fb_report":                                "googlefonts/repo/fb_report",
     "com.google.fonts/check/repo/sample_image":                             "googlefonts/repo/sample_image",
@@ -273,7 +274,6 @@ renaming_map = {
     "com.google.fonts/check/outline_jaggy_segments":                        "outline_jaggy_segments",
     "com.google.fonts/check/outline_semi_vertical":                         "outline_semi_vertical",
     "com.google.fonts/check/outline_short_segments":                        "outline_short_segments",
-    "com.google.fonts/check/render_own_name":                               "render_own_name",
     "com.google.fonts/check/required_tables":                               "required_tables",
     "com.google.fonts/check/rupee":                                         "rupee",
     "com.adobe.fonts/check/sfnt_version":                                   "sfnt_version",
