+++ Lib/fontbakery/profiles/fontbureau.py
@@ -25,7 +25,6 @@ PROFILE = {
         "nested_components",
         "no_mac_entries",
         "overlapping_path_segments",
-        "render_own_name",
         "smart_dropout",
         "stylisticset_description",
         "typographic_family_name",
