+++ tests/test_checks_googlefonts_render_own_name.py
@@ -1,0 +1,18 @@
+from conftest import check_id
+from fontbakery.codetesting import (
+    assert_PASS,
+    assert_results_contain,
+    TEST_FILE,
+)
+from fontbakery.status import FAIL
+
+
+@check_id("googlefonts/render_own_name")
+def test_check_render_own_name(check):
+    """Check family directory name."""
+
+    ttFont = TEST_FILE("overpassmono/OverpassMono-Regular.ttf")
+    assert_PASS(check(ttFont))
+
+    ttFont = TEST_FILE("noto_sans_tamil_supplement/NotoSansTamilSupplement-Regular.ttf")
+    assert_results_contain(check(ttFont), FAIL, "render-own-name")
